+++ include/libnavigate/gsv.h
@@ -65,7 +65,7 @@ NAVI_EXTERN(navierr_status_t) navi_init_gsv(struct gsv_t *msg, navi_talkerid_t t
 //
 // Creates GSV message
 NAVI_EXTERN(navierr_status_t) navi_create_gsv(const struct gsv_t *msg, char *buffer,
-	int maxsize, int *nmwritten);
+	size_t maxsize, size_t *nmwritten);
 
 //
 // Parses GSV message
